making spark test files ephemeral
avoid saving them to HDFS and regenerate per test suite
